+++ common/HAL/ports/STM32G4/Board.cpp
@@ -16,7 +16,7 @@ void Board::IO::initDrivers() {}
 
 void Board::IO::deinitPWM(){}
 
-void Board::IO::setMotorDutyCycle(enum motor motor, uint16_t duty_cycle){}
+void Board::IO::setMotorDutyCycle(enum motor motor, float duty_cycle){}
 
 int16_t Board::IO::getEncoderCount(enum encoder encoder){}
